Significant changes in many different aspects of the application. A key feature though is now I can toggle between UI and viewport. Will need to reimplement this properly in the future.

+++ src/utility.cpp
@@ -11,12 +11,12 @@ std::string load_text_file(std::string_view path)
 }
 
 
-float get_delta_time()
+double get_delta_time()
 {
     // todo: replace glfwGetTime() with C++ chrono
-    static float lastTime;
-    float currentTime = (float)glfwGetTime();
-    float deltaTime = currentTime - lastTime;
+    static double lastTime;
+    double currentTime = glfwGetTime();
+    double deltaTime = currentTime - lastTime;
     lastTime = currentTime;
 
     return deltaTime;
